flora-ml-api/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
databases
pydantic
PyJWT
passlib
loguru
starlette
apscheduler
torch
joblib
numpy
fastapi-cors
websockets
pandas
locust
redis
aiofiles
python-dotenv
numpy==1.26.4
onnx==1.16.1
onnxruntime-training-cpu==1.18.0
